# src/Cluely.Api/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Cluely API",
    "description": "REST API for Cluely authentication, room management, gameplay, and the Content Platform (dictionaries, discovery, sharing, and moderation).",
    "version": "v1"
  },
  "paths": {
...
    "/api/content": {
      "post": {
        "tags": [
          "Content"
        ],
        "summary": "Creates a new dictionary owned by the authenticated user.",
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Optional UUID. Reuse the same key to replay this command without creating a duplicate.",
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "7e7d0195-1d35-4ec8-b6d5-8d79fb77b509"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cluely.Api.Contracts.Requests.CreateContentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Cluely.Api.Contracts.Requests.CreateContentRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Cluely.Api.Contracts.Requests.CreateContentRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cluely.Api.Contracts.Responses.ContentCreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/content/{id}": {
      "patch": {
        "tags": [
          "Content"
        ],
        "summary": "Updates a dictionary's metadata.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cluely.Api.Contracts.Requests.UpdateContentRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Cluely.Api.Contracts.Requests.UpdateContentRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Cluely.Api.Contracts.Requests.UpdateContentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cluely.Api.Contracts.Responses.ContentUpdatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Content"
        ],
        "summary": "Archives a dictionary (soft delete); it can be restored.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "get": {
        "tags": [
          "ContentDiscovery"
        ],
        "summary": "Returns a dictionary's details, if the caller is permitted to see it.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cluely.Api.Contracts.Responses.ContentDetailsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/content/{id}/restore": {
      "post": {
        "tags": [
          "Content"
        ],
        "summary": "Restores a previously archived dictionary.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/content/{id}/words": {
      "post": {
        "tags": [
          "Content"
        ],
        "summary": "Adds words to the dictionary's draft.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cluely.Api.Contracts.Requests.AddWordsRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Cluely.Api.Contracts.Requests.AddWordsRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Cluely.Api.Contracts.Requests.AddWordsRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cluely.Api.Contracts.Responses.WordCountResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      },
      "patch": {
        "tags": [
          "Content"
        ],
        "summary": "Replaces a word in the dictionary's draft.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Cluely.Api.Contracts.Requests.ReplaceWordRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/Cluely.Api.Contracts.Requests.ReplaceWordRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/Cluely.Api.Contracts.Requests.ReplaceWordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cluely.Api.Contracts.Responses.WordCountResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/content/{id}/words/{word}": {
      "delete": {
        "tags": [
          "Content"
        ],
        "summary": "Removes a word from the dictionary's draft.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "word",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cluely.Api.Contracts.Responses.WordCountResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/content/{id}/validate": {
      "post": {
        "tags": [
          "Content"
        ],
        "summary": "Validates the dictionary's draft without publishing.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cluely.Api.Contracts.Responses.ValidateContentResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
    "/api/content/{id}/publish": {
      "post": {
        "tags": [
          "Content"
        ],
        "summary": "Publishes the dictionary's draft into a new immutable version.\r\nRepeating the request with the same Idempotency-Key returns the original version.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "description": "Required UUID. Retrying with the same key returns the original published version.",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "example": "7e7d0195-1d35-4ec8-b6d5-8d79fb77b509"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Cluely.Api.Contracts.Responses.PublishContentResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Microsoft.AspNetCore.Mvc.ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Authentication is required."
          }
        },
        "security": [
          {
            "Bearer": [ ]
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "Cluely.Api.Contracts.Requests.AddWordsRequest": {
        "required": [
          "words"
        ],
        "type": "object",
        "properties": {
          "words": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Words to add."
          }
        },
        "additionalProperties": false,
        "description": "Request to add words to a dictionary's draft."
      },
...
      "Cluely.Api.Contracts.Requests.CreateContentRequest": {
        "required": [
          "contentType",
          "language",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 256,
            "minLength": 1,
            "type": "string",
            "description": "Human-readable title."
          },
          "description": {
            "type": "string",
            "description": "Optional description.",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional tags for discovery.",
            "nullable": true
          },
          "language": {
            "minLength": 1,
            "type": "string",
            "description": "Intended audience language (advisory)."
          },
          "region": {
            "type": "string",
            "description": "Optional cultural region.",
            "nullable": true
          },
          "contentType": {
            "minLength": 1,
            "type": "string",
            "description": "Content type (currently \"user\")."
          }
        },
        "additionalProperties": false,
        "description": "Request to create a new dictionary (owned by the authenticated user)."
      },
...
      "Cluely.Api.Contracts.Requests.ReplaceWordRequest": {
        "required": [
          "existingWord",
          "newWord"
        ],
        "type": "object",
        "properties": {
          "existingWord": {
            "minLength": 1,
            "type": "string",
            "description": "The existing (normalized) word to replace."
          },
          "newWord": {
            "minLength": 1,
            "type": "string",
            "description": "The replacement word."
          }
        },
        "additionalProperties": false,
        "description": "Request to replace a word in a dictionary's draft."
      },
...
      "Cluely.Api.Contracts.Requests.UpdateContentRequest": {
        "required": [
          "language",
          "title"
        ],
        "type": "object",
        "properties": {
          "title": {
            "maxLength": 256,
            "minLength": 1,
            "type": "string",
            "description": "Human-readable title."
          },
          "description": {
            "type": "string",
            "description": "Optional description.",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Optional tags for discovery.",
            "nullable": true
          },
          "language": {
            "minLength": 1,
            "type": "string",
            "description": "Intended audience language (advisory)."
          },
          "region": {
            "type": "string",
            "description": "Optional cultural region.",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Request to update a dictionary's metadata."
      },
...
      "Cluely.Api.Contracts.Responses.ContentCreatedResponse": {
        "type": "object",
        "properties": {
          "dictionaryId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Response returned when a dictionary is created."
      },
      "Cluely.Api.Contracts.Responses.ContentDetailsResponse": {
        "type": "object",
        "properties": {
          "dictionaryId": {
            "type": "string",
            "format": "uuid"
          },
          "ownerId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "language": {
            "type": "string",
            "nullable": true
          },
          "region": {
            "type": "string",
            "nullable": true
          },
          "visibility": {
            "type": "string",
            "nullable": true
          },
          "contentType": {
            "type": "string",
            "nullable": true
          },
          "currentVersionId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "currentVersionLabel": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "versions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Cluely.Api.Contracts.Responses.ContentVersionResponse"
            },
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Detail view of a dictionary and the versions the caller may see."
      },
...
      "Cluely.Api.Contracts.Responses.ContentUpdatedResponse": {
        "type": "object",
        "properties": {
          "dictionaryId": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Response returned when a dictionary's metadata is updated."
      },
...
      "Cluely.Api.Contracts.Responses.WordCountResponse": {
        "type": "object",
        "properties": {
          "dictionaryId": {
            "type": "string",
            "format": "uuid"
          },
          "wordCount": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false,
        "description": "Response returned after a draft word-count change."
      },
      "Microsoft.AspNetCore.Mvc.ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}